+++ internal/crossdock/client/googlegrpcclient/googlegrpcclient.go
@@ -18,7 +18,7 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 // THE SOFTWARE.
 
-package googlegrpc
+package googlegrpcclient
 
 import (
 	"context"
