+++ internal/crossdock/client/start.go
@@ -27,7 +27,7 @@ import (
 	"go.uber.org/yarpc/internal/crossdock/client/errorshttpclient"
 	"go.uber.org/yarpc/internal/crossdock/client/errorstchclient"
 	"go.uber.org/yarpc/internal/crossdock/client/gauntlet"
-	"go.uber.org/yarpc/internal/crossdock/client/googlegrpc"
+	"go.uber.org/yarpc/internal/crossdock/client/googlegrpcclient"
 	"go.uber.org/yarpc/internal/crossdock/client/grpc"
 	"go.uber.org/yarpc/internal/crossdock/client/headers"
 	"go.uber.org/yarpc/internal/crossdock/client/httpserver"
@@ -45,7 +45,7 @@ var behaviors = crossdock.Behaviors{
 	"json":                  echo.JSON,
 	"thrift":                echo.Thrift,
 	"protobuf":              echo.Protobuf,
-	"google_grpc_client":    googlegrpc.Run,
+	"google_grpc_client":    googlegrpcclient.Run,
 	"grpc":                  grpc.Run,
 	"headers":               headers.Run,
 	"errors_httpclient":     errorshttpclient.Run,
